Willy2

+++ src/main/java/com/example/sistematransmilenio/service/BusService.java
@@ -2,6 +2,8 @@ package com.example.sistematransmilenio.service;
 
 import com.example.sistematransmilenio.model.Bus;
 import com.example.sistematransmilenio.model.Conductor;
+import com.example.sistematransmilenio.model.Horario;
+import com.example.sistematransmilenio.repository.HorarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.example.sistematransmilenio.repository.BusRepository;
@@ -12,6 +14,9 @@ import java.util.List;
 public class BusService {
 
     @Autowired
+    private HorarioRepository horarioRepository;
+
+    @Autowired
     private BusRepository busRepository;
     public List<Bus> listarBuses (){return busRepository.findAll();}
 
@@ -20,14 +25,9 @@ public class BusService {
 
     public Bus findBusById(Long id){return busRepository.findById(id).get();}
 
-
-
-
     public Bus update(Bus bus){
         return busRepository.save(bus);
     }
-
-    public Bus findBusByPlaca(String placa){return busRepository.findBusByPlaca(placa);}
 
     public void guardarBus(Bus bus) {
         busRepository.save(bus);
@@ -41,8 +41,34 @@ public class BusService {
         return busRepository.findBusByPlacaStartingWith(textoBusqueda);
     }
 
-    public void eliminarBus(long id ){
-        busRepository.deleteById(id);
+    public boolean eliminarBus(long id ){
+
+        if(this.validarBus(id)){
+            try{
+                busRepository.deleteById(id);
+                return true;
+            }catch(Exception e){
+                return false;
+            }
+        }
+
+        return true;
+    }
+    private boolean validarBus(Long id) {
+        List<Horario> horarios = horarioRepository.getBusByConductorId(id);
+        if(horarios.isEmpty()){
+            return true;
+        }else{
+            for(Horario h: horarios){
+                System.out.println(h.toString());
+            }
+
+            return false;
+        }
+    }
+
+    public Bus findBusByPlaca(String placa) {
+        return this.busRepository.findBusByPlaca(placa);
     }
 }
